scroll mouse added to files
scroll mouse feature was added to the habits, tasks, dailies, shop, and
list scroll boxes. Note: scrolling only works if the mouse is directly
on the canvas and not on any frames or labels inside the canvas.
Bind_all solves this problem but I couldn't get multiple instances of
the MouseWheel event to work for more than one canvas when using the
bind_all function...