Tool tips active, inventory bugs fixed, and more.
Tool tips are set to 0.25 second delays on active items, shop items, and
inventory items.
To see the tool tip, the mouse can NOT move or else the delay will
reset.

There were some inventory errors with the item count and they are now
fixed. I re-styled
the shop slightly too.

Also, a nice feature I added to both the shop and inventory, is the
ability to use the mouse
wheel to scroll. This worked before, but now the user can be hovering
over an item and the
mouse wheel will still scroll the canvas.

I took out all the white space on all items and deleted png files since
they are not being used.
I started working on item emblems for the active item interface, so not
all of them are finished.