Boss GUI placement & main char emotions
The first boss is displayed graphically and some basic info on the boss
such as health and distance. The main character image changes to "happy"
when a hack is completed and "mad" when a hack is deleted. Also, some
images were not sized properly, so I resized them and changed the names
on the bosses as well.